+++ tests/test_parser_record_breakend.py
@@ -31,7 +31,7 @@ MEDIUM_HEADER = """
 ##FORMAT=<ID=GQ,Number=1,Type=Integer,Description="Genotype Quality">
 ##FORMAT=<ID=DP,Number=1,Type=Integer,Description="Read Depth">
 ##FORMAT=<ID=HQ,Number=2,Type=Integer,Description="Haplotype Quality">
-##FORMAT=<ID=FT,Number=.,Type=String,Description="Call-wise filters">
+##FORMAT=<ID=FT,Number=1,Type=String,Description="Call-wise filters">
 #CHROM\tPOS\tID\tREF\tALT\tQUAL\tFILTER\tINFO\tFORMAT\tNA00001\tNA00002\tNA00003
 """.lstrip()